+++ compact_str/src/repr/heap.rs
@@ -1,20 +1,10 @@
-use core::mem;
-
 // use super::arc::ArcString;
 use super::boxed::BoxString;
-use super::{
-    HEAP_MASK,
-    MAX_SIZE,
-};
-
-const PADDING_SIZE: usize = MAX_SIZE - mem::size_of::<BoxString>();
-const PADDING: [u8; PADDING_SIZE] = [HEAP_MASK; PADDING_SIZE];
 
 #[repr(C)]
 #[derive(Debug, Clone)]
 pub struct HeapString {
     pub string: BoxString,
-    padding: [u8; PADDING_SIZE],
 }
 
 impl HeapString {
@@ -24,29 +14,35 @@ impl HeapString {
     /// `HeapString::with_additional()`
     #[inline]
     pub fn new(text: &str) -> Self {
-        let padding = PADDING;
         let string = BoxString::new(text);
-
-        HeapString { padding, string }
+        HeapString { string }
     }
 
     /// Creates a [`HeapString`] from the provided `text` and allocates the underlying buffer with
     /// `additional` capacity
     #[inline]
     pub fn with_additional(text: &str, additional: usize) -> Self {
-        let padding = PADDING;
         let string = BoxString::with_additional(text, additional);
-
-        HeapString { padding, string }
+        HeapString { string }
     }
 
     /// Creates a [`HeapString`] with the provided capacity.
     #[inline]
     pub fn with_capacity(capacity: usize) -> Self {
-        let padding = PADDING;
         let string = BoxString::with_capacity(capacity);
+        HeapString { string }
+    }
 
-        HeapString { padding, string }
+    #[inline]
+    pub fn from_string(s: String) -> Self {
+        let string = BoxString::from_string(s);
+        HeapString { string }
+    }
+
+    #[inline]
+    pub fn from_box_str(b: Box<str>) -> Self {
+        let string = BoxString::from_box_str(b);
+        HeapString { string }
     }
 
     /// Makes a mutable reference to the underlying buffer.
@@ -61,16 +57,6 @@ impl HeapString {
     #[inline]
     pub unsafe fn set_len(&mut self, length: usize) {
         self.string.set_len(length)
-    }
-}
-
-impl From<String> for HeapString {
-    #[inline]
-    fn from(s: String) -> Self {
-        let padding = PADDING;
-        let string = BoxString::from(s.as_str());
-
-        HeapString { padding, string }
     }
 }