+++ fuzz/src/lib.rs
@@ -18,12 +18,22 @@ pub struct Scenario<'a> {
 
 #[derive(Arbitrary, Debug)]
 pub enum Creation<'a> {
+    /// Create using [`CompactStr::from_utf8`]
     Bytes(&'a [u8]),
+    /// Create using [`CompactStr::from_utf8_buf`]
     Buf(&'a [u8]),
+    /// Create using an iterator of chars (i.e. the `FromIterator` trait)
     IterChar(Vec<char>),
+    /// Create using an iterator of strings (i.e. the `FromIterator` trait)
     IterString(Vec<String>),
+    /// Create using [`CompactStr::new`]
     Word(String),
+    /// Create using [`CompactStr::from_utf8_buf`] when the buffer is non-contiguous
     NonContiguousBuf(&'a [u8]),
+    /// Create using `From<String>`, which consumes the `String` for `O(1)` runtime
+    FromString(String),
+    /// Create using `From<Box<str>>`, which consumes the `Box<str>` for `O(1)` runtime
+    FromBoxStr(Box<str>),
 }
 
 impl Creation<'_> {
@@ -31,7 +41,6 @@ impl Creation<'_> {
         use Creation::*;
 
         match self {
-            // Create a `CompactStr` from a `String`
             Word(word) => {
                 let compact = CompactStr::new(&word);
 
@@ -40,7 +49,37 @@ impl Creation<'_> {
 
                 Some((compact, word))
             }
-            // Create a `CompactStr` from an iterator of `char`s
+            FromString(s) => {
+                let compact = CompactStr::from(s.clone());
+
+                assert_eq!(compact, s);
+
+                // Note: converting From<String> will always be heap allocated because we use the
+                // underlying buffer from the source String
+                if s.capacity() == 0 {
+                    assert!(!compact.is_heap_allocated());
+                } else {
+                    assert!(compact.is_heap_allocated());
+                }
+
+                Some((compact, s))
+            }
+            FromBoxStr(b) => {
+                let compact = CompactStr::from(b.clone());
+
+                assert_eq!(compact, b);
+
+                // Note: converting From<Box<str>> will always be heap allocated because we use the
+                // underlying buffer from the source String
+                if b.len() == 0 {
+                    assert!(!compact.is_heap_allocated())
+                } else {
+                    assert!(compact.is_heap_allocated())
+                }
+
+                let string = String::from(b);
+                Some((compact, string))
+            }
             IterChar(chars) => {
                 let compact: CompactStr = chars.iter().collect();
                 let std_str: String = chars.iter().collect();
@@ -50,7 +89,6 @@ impl Creation<'_> {
 
                 Some((compact, std_str))
             }
-            // Create a `CompactStr` from an iterator of `String`s
             IterString(strings) => {
                 let compact: CompactStr = strings.iter().map::<&str, _>(|s| s.as_ref()).collect();
                 let std_str: String = strings.iter().map::<&str, _>(|s| s.as_ref()).collect();
@@ -60,7 +98,6 @@ impl Creation<'_> {
 
                 Some((compact, std_str))
             }
-            // Create a `CompactStr` from a slice of bytes
             Bytes(data) => {
                 let compact = CompactStr::from_utf8(data);
                 let std_str = std::str::from_utf8(data);
@@ -81,7 +118,6 @@ impl Creation<'_> {
                     _ => panic!("CompactStr and core::str read UTF-8 differently?"),
                 }
             }
-            // Create a `CompactStr` from a buffer of bytes
             Buf(data) => {
                 let mut buffer = Cursor::new(data);
 
@@ -104,7 +140,6 @@ impl Creation<'_> {
                     _ => panic!("CompactStr and core::str read UTF-8 differently?"),
                 }
             }
-            // Create a `CompactStr` from a non-contiguous buffer of bytes
             NonContiguousBuf(data) => {
                 let mut queue = if data.len() > 3 {
                     // if our data is long, make it non-contiguous